переносит методы с фильтрацией и сортировкой из main.js в filters.js
